+++ noodles-fasta/src/fai/async/reader.rs
@@ -1,4 +1,4 @@
-use tokio::io::{self, AsyncBufRead, AsyncBufReadExt};
+use tokio::io::{self, AsyncBufRead};
 
 use crate::fai::Index;
 
@@ -50,6 +50,8 @@ where
     /// # }
     /// ```
     pub async fn read_index(&mut self) -> io::Result<Index> {
+        use crate::r#async::reader::read_line;
+
         let mut buf = String::new();
         let mut index = Vec::new();
 
@@ -69,51 +71,5 @@ where
         }
 
         Ok(index)
-    }
-}
-
-async fn read_line<R>(reader: &mut R, buf: &mut String) -> io::Result<usize>
-where
-    R: AsyncBufRead + Unpin,
-{
-    const LINE_FEED: char = '\n';
-    const CARRIAGE_RETURN: char = '\r';
-
-    match reader.read_line(buf).await? {
-        0 => Ok(0),
-        n => {
-            if buf.ends_with(LINE_FEED) {
-                buf.pop();
-
-                if buf.ends_with(CARRIAGE_RETURN) {
-                    buf.pop();
-                }
-            }
-
-            Ok(n)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_read_line() -> io::Result<()> {
-        async fn t(buf: &mut String, mut data: &[u8], expected: &str) -> io::Result<()> {
-            buf.clear();
-            read_line(&mut data, buf).await?;
-            assert_eq!(buf, expected);
-            Ok(())
-        }
-
-        let mut buf = String::new();
-
-        t(&mut buf, b"noodles\n", "noodles").await?;
-        t(&mut buf, b"noodles\r\n", "noodles").await?;
-        t(&mut buf, b"noodles", "noodles").await?;
-
-        Ok(())
     }
 }